+++ sm/tedge_agent/src/main.rs
@@ -1,8 +1,4 @@
 use agent::SmAgentConfig;
-use tedge_config::{
-    ConfigRepository, ConfigSettingAccessor, ConfigSettingAccessorStringExt, MqttPortSetting,
-    SoftwarePluginDefaultSetting,
-};
 use tedge_users::UserManager;
 
 mod agent;
@@ -11,28 +7,13 @@ mod state;
 
 #[tokio::main]
 async fn main() -> Result<(), anyhow::Error> {
-    let user_manager = UserManager::new();
-
-    let tedge_config_location = tedge_config::TEdgeConfigLocation::from_default_system_location();
-    let config_repository = tedge_config::TEdgeConfigRepository::new(tedge_config_location);
-    let tedge_config = config_repository.load()?;
-
     initialise_logging();
-
-    let default_plugin_type = tedge_config.query_string_optional(SoftwarePluginDefaultSetting)?;
-    let mqtt_config =
-        mqtt_client::Config::default().with_port(tedge_config.query(MqttPortSetting)?.into());
-
-    let tedge_config_path = config_repository
-        .get_config_location()
-        .tedge_config_root_path()
-        .to_path_buf();
-    let sm_agent_config = SmAgentConfig::default()
-        .with_default_plugin_type(default_plugin_type)
-        .with_sm_home(tedge_config_path)
-        .with_mqtt_client_config(mqtt_config);
-
-    let agent = agent::SmAgent::new("tedge_agent", sm_agent_config, user_manager);
+    let tedge_config_location = tedge_config::TEdgeConfigLocation::from_default_system_location();
+    let agent = agent::SmAgent::new(
+        "tedge_agent",
+        SmAgentConfig::try_new(tedge_config_location)?,
+        UserManager::new(),
+    );
     agent.start().await?;
     Ok(())
 }
